created pages related to terms and conditions
I created pages related to terms and conditions and added content for them.
I do not like how the layout for "subprocessors" looks. Compare it to the active enlist page. I could not figure out how to align the text like it is on the live site using markdown.
Added gdpr link within the misc/...small businesses....md content
Made an edit to the defult layout for a single page for when it calls the partial jumbotron.