Mobile Badges
I am still getting my butt kicked by these badges. I downloaded the "official" badges. I was able to remove an invisiable background from the Google badge that was making it very difficult to resize to the size of the apple badge. The badges are now very close to the same height but are slightly off. I resized the Google badge to the same px height as the Apple one and then moved the file into hugo but it still looks off. I managed to get the badges to look good on mobile view when I chrome inspect. It still really bothers me that I can't figure out how to keep the images from moving when you resize the browser. The End.