navigation,css,and content
I modified css on multiple files.
I added navigation for /security, /data-processing-agreement, and / subprocessors to the /gdpr.md
I removed some unused lines of code and markdown in certain files
I deleted the Values folder from content. Values was a page that came with the template.
I made a shortcode that displays a featured banner for resources page
Sorry for all the changes in this PR I got carried away.