improved on the leaf bundles and made some changes to layouts
I removed the description and image tags from the front matter of all articles in resources. They now use the img short code for the responsive image and .Summary for the description. I deleted the responsive image code from the single.html layout for resources and made another short code to hold it incase we want to call a responsive image in front matter somewhere else. I also deleted most of the image content that came with the hugo theme we are using.